pyproject.toml
[project]
name = "deep-finance-research-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "beautifulsoup4>=4.14.3",
    "datasets>=4.5.0",
    "ddgs>=9.10.0",
    "deepagents>=0.3.9",
    "docling[vlm]>=2.71.0",
    "docutils>=0.22.4",
    "docx2txt>=0.9",
    "duckduckgo-search>=8.1.1",
    "email-validator>=2.3.0",
    "evaluate>=0.4.6",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "fastembed>=0.7.4",
    "fastembed-gpu>=0.7.4",
    "google-genai>=1.61.0",
    "hf-xet>=1.2.0",
    "huggingface-hub>=0.36.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-google-vertexai>=3.2.1",
    "langchain-huggingface>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.7",
    "langgraph-api>=0.7.13",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.12",
    "langsmith>=0.6.7",
    "lxml>=6.0.2",
    "markdown2>=2.5.4",
    "markitdown[all]>=0.1.4",
    "nest-asyncio>=1.6.0",
    "nltk>=3.9.2",
    "numba>=0.62.1",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "openpyxl>=3.1.5",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pymupdf4llm>=0.2.9",
    "pymysql>=1.1.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-pptx>=1.0.2",
    "pytube>=15.0.0",
    "rank-bm25>=0.2.2",
    "selenium>=4.32.0",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "tavily-python>=0.7.21",
    "textblob>=0.19.0",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.2",
    "torch>=2.10.0",
    "transformers>=4.57.6",
    "typing-extensions>=4.15.0",
    "unstructured>=0.18.31",
    "wikipedia>=1.4.0",
    "xhtml2pdf>=0.2.17",
    "xmltodict>=1.0.2",
    "youtube-transcript-api>=1.0.3",
]
